Remove CMDOptions and ViewSpecification.
This change removes CMDOptions and ViewSpecification and replaces them with
PrintSettings. It also removes dead code that was only reachable with old
settings that no longer have command line options.

The aim is to:
- Clean up and simplify how the settings are set by command line options.
- Reduce and modernize the code.
- Work towards using DIVA as a library.
- Make it easier to remove global state in the future.

There are some things in PrintSettings that should be seperated out later as
they aren't tightly related to the content/format of the printed output. Doing
so will be part of the next step, which is to seperate the printing settings
from the Reader and 'de-global' them.